remove `NavigationTree` widget
