+++ packages/widgetbook/lib/src/navigation/widgets/widgets.dart
@@ -1,5 +1,4 @@
 export 'navigation_panel.dart';
-export 'navigation_tree.dart';
 export 'navigation_tree_item.dart';
 export 'navigation_tree_node.dart';
 export 'search_field.dart';
